# src/main/java/com/karumien/hris/openapi/server/controller/CustomApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.2.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.karumien.hris.openapi.server.controller;

import javax.validation.Valid;

import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import com.karumien.hris.openapi.model.hris.CustomRequest;
import com.karumien.hris.openapi.model.hris.CustomResponse;
import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;


@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Api(value = "custom", description = "the custom API")
public interface CustomApi {

    /**
     * GET /custom : Accepts request, replace it with mapper and process to Kafka
     * Accepts request, replace it with mapper and process to Kafka
     *
     * @param customRequest (required)
     * @return Model without API but paths need to be present in order to generate (status code 200)
     */
    @ApiOperation(value = "Accepts request, replace it with mapper and process to Kafka", nickname = "replace", notes = "Accepts request, "
            + "replace it with mapper and process to Kafka", response = CustomResponse.class, tags = {"replace",})
    @ApiResponses(value = {
            @ApiResponse(code = 200, message = "Model without API but paths need to be present in order to generate", response =
                    CustomResponse.class)
    })
    @RequestMapping(
            method = RequestMethod.GET,
            value = "/custom",
            produces = {"application/json"},
            consumes = {"application/json"}
    )
    ResponseEntity<CustomResponse> replace(@ApiParam(value = "", required = true) @Valid @RequestBody CustomRequest customRequest);

}